:rotating_light: Fix implicit cast

+++ Source/PluginEditor_.cpp
@@ -26,8 +26,8 @@ void PolesAndZerosEQAudioProcessorEditor::sizeSetup()
     applicationProperties.setStorageParameters(options);
     if (juce::PropertiesFile* pf
         = applicationProperties.getCommonSettings(true))
-        claf.setResizeRatio(
-            pf->getDoubleValue("sizeRatio", claf.getResizeRatio()));
+        claf.setResizeRatio(static_cast<float>(
+            pf->getDoubleValue("sizeRatio", claf.getResizeRatio())));
 
     setResizable(true, true);
     setSize(claf.getResizedWidth(), claf.getResizedHeight());
